숙박 검색 및 리뷰 리팩토링

+++ src/main/java/com/mip/sharebnb/service/ReviewService.java
@@ -3,28 +3,25 @@ package com.mip.sharebnb.service;
 import com.mip.sharebnb.dto.ReviewDto;
 import com.mip.sharebnb.exception.DataNotFoundException;
 import com.mip.sharebnb.exception.DuplicateValueExeption;
+import com.mip.sharebnb.exception.InvalidInputException;
 import com.mip.sharebnb.model.Accommodation;
 import com.mip.sharebnb.model.Member;
 import com.mip.sharebnb.model.Reservation;
 import com.mip.sharebnb.model.Review;
-import com.mip.sharebnb.repository.AccommodationRepository;
-import com.mip.sharebnb.repository.MemberRepository;
 import com.mip.sharebnb.repository.ReservationRepository;
 import com.mip.sharebnb.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
-
-    private final AccommodationRepository accommodationRepository;
-
-    private final MemberRepository memberRepository;
 
     private final ReservationRepository reservationRepository;
 
@@ -36,20 +33,40 @@ public class ReviewService {
 
     public void writeReview(ReviewDto reviewDto) {
 
-        if (reviewRepository.findReviewByReservationId(reviewDto.getReservationId()).isPresent()) {
-            throw new DuplicateValueExeption("Already Have a Review");
-        }
-
         Reservation reservation = reservationRepository.findById(reviewDto.getReservationId())
                 .orElseThrow(() -> new DataNotFoundException("Reservation Not Found"));
 
-        Accommodation accommodation = accommodationRepository.findById(reviewDto.getAccommodationId())
-                .orElseThrow(() -> new DataNotFoundException("Accommodation Not Found"));
+        if (reviewRepository.findReviewByReservationId(reviewDto.getReservationId()).isPresent()
+                || reservation.getIsWrittenReview()) {
+            throw new DuplicateValueExeption("Already Have a Review");
+        }
 
-        Member member = memberRepository.findById(reviewDto.getMemberId())
-                .orElseThrow(() -> new DataNotFoundException("Member Not Found"));
+        Accommodation accommodation = reservation.getAccommodation();
+        if (accommodation == null) {
+            throw new DataNotFoundException("Accommodation Not Found");
+        }
+
+        if (accommodation.getId() != reviewDto.getAccommodationId()) {
+            throw new InvalidInputException("Accommodation Not Matched");
+        }
+
+        Member member = reservation.getMember();
+        if (member == null) {
+            throw new DataNotFoundException("Member Not Found");
+        }
+
+        if (member.getId() != reviewDto.getMemberId()) {
+            throw new InvalidInputException("Member Not Matched");
+        }
+
+        int newReviewNum = accommodation.getReviewNum() + 1;
+        float newRating = (accommodation.getRating() * accommodation.getReviewNum() + reviewDto.getRating())
+                / newReviewNum;
 
         reservation.setIsWrittenReview(true);
+        accommodation.setRating(newRating);
+        accommodation.setReviewNum(newReviewNum);
+        accommodation.setHostReviewNum(accommodation.getHostReviewNum() + 1);
 
         Review review = new Review();
         review.setRating(reviewDto.getRating());
@@ -66,18 +83,57 @@ public class ReviewService {
                 .findReviewByReservationId(reviewDto.getReservationId())
                 .orElseThrow(() -> new DataNotFoundException("Review Not Found"));
 
+        Reservation reservation = reservationRepository.findById(reviewDto.getReservationId())
+                .orElseThrow(() -> new DataNotFoundException("Reservation Not Found"));
+
+        Accommodation accommodation = reservation.getAccommodation();
+        if (accommodation == null) {
+            throw new DataNotFoundException("Accommodation Not Found");
+        }
+
+        if (accommodation.getId() != reviewDto.getAccommodationId()) {
+            throw new InvalidInputException("Accommodation Not Matched");
+        }
+
+        Member member = reservation.getMember();
+        if (member == null) {
+            throw new DataNotFoundException("Member Not Found");
+        }
+
+        if (member.getId() != reviewDto.getMemberId()) {
+            throw new InvalidInputException("Member Not Matched");
+        }
+
+        float originRating = accommodation.getRating();
+        float newRating = originRating + (reviewDto.getRating() - originReview.getRating()) / accommodation.getReviewNum();
+        accommodation.setRating(newRating);
+
+        originReview.setAccommodation(accommodation);
         originReview.setContent(reviewDto.getContent());
         originReview.setRating(reviewDto.getRating());
         originReview.setCreatedDate(LocalDate.now());
-
-        reviewRepository.save(originReview);
     }
 
     public void deleteReview(Long id) {
-        Review originReview = reviewRepository.findById(id)
+        Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("Review Not Found"));
 
-        originReview.getReservation().setIsWrittenReview(false);
+        Reservation reservation = review.getReservation();
+        if (reservation == null) {
+            throw new DataNotFoundException("Reservation Not Found");
+        }
+        reservation.setIsWrittenReview(false);
+
+
+        Accommodation accommodation = reservation.getAccommodation();
+        if (accommodation == null) {
+            throw new DataNotFoundException("Accommodation Not Found");
+        }
+
+        accommodation.setRating((accommodation.getRating() * accommodation.getReviewNum() - review.getRating())
+                / (accommodation.getReviewNum() - 1));
+        accommodation.setReviewNum(accommodation.getReviewNum() - 1);
+        accommodation.setHostReviewNum(accommodation.getHostReviewNum() - 1);
 
         reviewRepository.deleteById(id);
     }
